# openapi.yaml
openapi: 3.0.3
info:
  title: ADADProject API
  version: "1.0.0"
  description: |
    REST API for ADADProject (events and users). This document describes the endpoints implemented in `routes/events.js` and `routes/users.js`.
servers:
  - url: http://localhost:3000
paths:
  /events:
    get:
      summary: List events (paginated)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
      responses:
        '200':
          description: Paginated list of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
    post:
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  acknowledged:
                    type: boolean
                  insertedId:
                    type: string
  /events/top/{limit}:
    get:
      summary: Top events by average score
      parameters:
        - in: path
          name: limit
          required: true
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Events ordered by average score desc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventWithStats'
  /events/ratings/{order}:
    get:
      summary: Events ordered by number of reviews
      parameters:
        - in: path
          name: order
          required: true
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Events ordered by review count
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventWithReviewCount'
...
  /events/county/{county}:
    get:
      summary: Events in a county with stats
      parameters:
        - in: path
          name: county
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Events in county with averageScore and reviewsCount
          content:
            application/json:
              schema:
                type: object
                properties:
                  county:
                    type: string
                  totalEvents:
                    type: integer
                  countyAverage:
                    type: number
                    nullable: true
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventWithStats'
  /events/{idOrYear}:
    get:
      summary: Get event by id or list events reviewed in a year
      description: |
        This path accepts either an event ObjectId string or a 4-digit year. If an ObjectId is provided, the response is the event with averageScore and reviewsCount. If a year is provided, the response contains events reviewed in that year.
      parameters:
        - in: path
          name: idOrYear
          required: true
          schema:
            type: string
          description: ObjectId of an event or a 4-digit year (e.g. 2024)
      responses:
        '200':
          description: Event object (when id provided) or events by year (when year provided)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/EventWithStats'
                  - $ref: '#/components/schemas/EventsByYear'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /events/{id}:
    put:
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: Fields to update (partial allowed)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
      responses:
        '200':
          description: Updated event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted flag
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Event:
      type: object
      properties:
        _id:
          type: string
        changeDate:
          type: string
          description: original changeDate used by the project dataset
        establishmentID:
          type: string
        establishmentName:
          type: string
        address:
          type: string
        zipCode:
          type: string
        county:
          type: string
      example:
        _id: "650a3c..."
        changeDate: "2023-05-01T00:00:00.000Z"
        establishmentID: "E123"
        establishmentName: "Sample Place"
        address: "123 Main St"
        zipCode: "12345"
        county: "SomeCounty"
    EventCreate:
      type: object
      required: [changeDate, establishmentID, establishmentName, address, zipCode, county]
      properties:
        changeDate:
          type: string
          format: date-time
        establishmentID:
          type: string
        establishmentName:
          type: string
        address:
          type: string
        zipCode:
          type: string
        county:
          type: string
    EventUpdate:
      type: object
      properties:
        changeDate:
          type: string
        establishmentID:
          type: string
        establishmentName:
          type: string
        address:
          type: string
        zipCode:
          type: string
        county:
          type: string
...
    EventWithStats:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            averageScore:
              type: number
              nullable: true
            reviewsCount:
              type: integer
    EventWithReviewCount:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            reviewsCount:
              type: integer
...
    EventsByYear:
      type: object
      properties:
        year:
          type: integer
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
...
    PaginatedEvents:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
...